1. add model and controller for articles and separate out functionality from topics controller, 2. add getArticles function to controller, 3. add fetchArticles function to model
