+++ controllers/articles.controllers.js
@@ -1,0 +1,25 @@
+const {fetchArticleById, fetchArticles} = require("../models/articles.models")
+
+const getArticleByID = (req, res, next) => {
+    const {article_id} = req.params
+
+    fetchArticleById(article_id)
+    .then((article) => {
+        res.status(200).send({article})
+    })
+    .catch((err) => {
+        next(err)
+    })
+}
+
+const getArticles = (req, res, next) => {
+    fetchArticles()
+    .then((articles) => {
+        res.status(200).send({articles})
+    })
+    .catch((err) => {
+        next(err)
+    })
+}
+
+module.exports = {getArticleByID, getArticles}
