+++ controllers/topics.controllers.js
@@ -1,4 +1,4 @@
-const {fetchTopics, fetchArticleById} = require("../models/topics.models")
+const {fetchTopics} = require("../models/topics.models")
 
 
 const getTopics = (req, res, next) => {
@@ -11,21 +11,7 @@ const getTopics = (req, res, next) => {
     });
 }
 
-const getArticleByID = (req, res, next) => {
-    const {article_id} = req.params
-
-    fetchArticleById(article_id)
-    .then((article) => {
-        res.status(200).send({article})
-    })
-    .catch((err) => {
-        next(err)
-    })
-}
 
 
 
-
-//request.params will extract the article id number and that will be passed into the model- look at the notes 
-
-module.exports = {getTopics, getArticleByID}
+module.exports = {getTopics}
