+++ models/topics.models.js
@@ -9,21 +9,5 @@ const fetchTopics = () => {
 }
 
 
-const fetchArticleById = (article_id) => {
-    return db
-    .query('SELECT * FROM articles WHERE article_id = $1', [article_id])
-    .then(({rows}) => {
-        const article = rows[0]
-    if (!article) {
-        return Promise.reject({
-            status: 404,
-            msg: "No article associated with this id number",
-          })
-    }
-     return article
-    })
-}
-
-
-module.exports = {fetchTopics, fetchArticleById}
+module.exports = {fetchTopics}
